[DDW-421] Screenshot method importable
